% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ntc_app}
\alias{ntc_app}
\title{Run the NtC Shiny app}
\usage{
ntc_app(...)
}
\arguments{
\item{...}{passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}} options, e.g. \code{launch.browser}}
}
\value{
a shiny app object
}
\description{
Run the NtC Shiny app
}
