% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/original.R
\name{ntc_original}
\alias{ntc_original}
\title{Original NtC algorithm (Stadnicka-Michalak et al.)}
\usage{
ntc_original(data, measured_criterion = TRUE, narrow_ci_correction = TRUE)
}
\arguments{
\item{data}{an \code{ntc_data} object from \code{\link[=ntc_prepare]{ntc_prepare()}}}

\item{measured_criterion}{apply the measured value criterion}

\item{narrow_ci_correction}{apply the correction for narrow confidence intervals}
}
\value{
list with the NtC, intermediate values, the fit and the confidence band
}
\description{
Original NtC algorithm (Stadnicka-Michalak et al.)
}
