% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntc.R
\name{ntc_table}
\alias{ntc_table}
\title{Results as a table}
\usage{
ntc_table(result)
}
\arguments{
\item{result}{an \code{ntc_result}}
}
\value{
data frame with one row per quantity
}
\description{
Results as a table
}
