% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ntc}
\alias{plot_ntc}
\title{Plot the fitted curve, confidence band, data and NtC}
\usage{
plot_ntc(result, units = "units")
}
\arguments{
\item{result}{an \code{ntc_result}}

\item{units}{concentration units for the axis label}
}
\value{
a ggplot object
}
\description{
Plot the fitted curve, confidence band, data and NtC
}
